project/llm-worker-youtube/requirements.txt
openai==0.28
python-dotenv
grpcio==1.62.1
grpcio-tools==1.62.1
youtube-transcript-api==1.1.1